+++ amp/descriptor/analysis.py
@@ -24,6 +24,7 @@ class FingerprintPlot:
         from matplotlib import pyplot
         from matplotlib.backends.backend_pdf import PdfPages
 
+
         self.compile_fingerprints(images)
 
         self.figures = {}