+++ amp/model/neuralnetwork.py
@@ -1,15 +1,16 @@
 import os
 import numpy as np
 from collections import OrderedDict
+from scipy.optimize import fmin
 from ase.calculators.calculator import Parameters
 
 from . import LossFunction, calculate_fingerprints_range, Model
 from ..regression import Regressor
 from ..utilities import Logger, hash_images, make_filename
+from .. import Amp
 
 
 class NeuralNetwork(Model):
-
     """Class that implements a basic feed-forward neural network.
 
     Parameters
@@ -44,20 +45,20 @@ class NeuralNetwork(Model):
         function, "tanh" refers to tanh function, and "sigmoid" refers to
         sigmoid function.
     weights : dict
-        In the case of no descriptor, keys correspond to layers and values are
-        two dimensional arrays of network weight.  In the atom-centered mode,
-        keys correspond to chemical elements and values are dictionaries with
-        layer keys and network weight two dimensional arrays as values. Arrays
-        are set up to connect node i in the previous layer with node j in the
-        current layer with indices w[i,j]. The last value for index i
-        corresponds to bias. If weights is not given, arrays will be randomly
-        generated.
+        In the atom-centered mode, keys correspond to chemical elements and
+        values are dictionaries with layer keys and network weight two
+        dimensional arrays as values. Arrays are set up to connect node i in
+        the previous layer with node j in the current layer with indices
+        w[i,j]. The last value for index i corresponds to bias.  In the case of
+        no descriptor, keys correspond to layers and values are two dimensional
+        arrays of network weight.  If weights is not given, arrays will be
+        randomly generated.
     scalings : dict
         In the case of no descriptor, keys are "intercept" and "slope" and
         values are real numbers. In the fingerprinting scheme, keys correspond
         to chemical elements and values are dictionaries with "intercept" and
-        "slope" keys and real number values. If scalings is not given, it will
-        be randomly generated.
+        "slope" keys and real number values. If scalings is None, it will
+        be guessed based on the input range of energies.
     fprange : dict
         Range of fingerprints of each chemical species.  Should be fed as
         a dictionary of chemical species and a list of minimum and maximun,
@@ -65,25 +66,36 @@ class NeuralNetwork(Model):
 
         >>> fprange={"Pd": [0.31, 0.59], "O":[0.56, 0.72]}
 
+    mode : str
+        Can be either 'atom-centered' or 'image-centered'.
+    version : object
+        Version of this class.
     regressor : object
         Regressor object for finding best fit model parameters, e.g. by loss
         function optimization via amp.regression.Regressor.
-    mode : str
-        Can be either 'atom-centered' or 'image-centered'.
     lossfunction : object
         Loss function object.
-    retries : int
-        If model does not converge during training, number of times to retry
-        before giving up.
-    version : object
-        Version of this class.
+    prescale : bool
+        If True, will start with a simple single-parameter (per element)
+        regression to find the best-fit values for the scaling parameters. This
+        portion of the code is not parallelized, but can greatly improve the
+        likelihood of and path to convergence, as this is the most sensitive
+        parameter.  Ignored if scalings are explicitly supplied.
     fortran : bool
         Can optionally shut off fortran, primarily for debugging.
     checkpoints : int
         Frequency with which to save parameter checkpoints upon training. E.g.,
-        100 saves a checkpoint on each 100th training setp.  Specify None for
-        no checkpoints. Note: You can make this negative to not overwrite
-        previous checkpoints.
+        100 saves a checkpoint on each 100th training step.  By default only
+        the last checkpoint is retained; to keep all checkpoints make the
+        frequency negative instead of positive.  Specify None for no
+        checkpoints.  Note that checkpoints can be used to resume a training
+        run simply by resubmitting the same script; if a checkpoint file is
+        found it will be used.
+    randomseed : None or int
+        Seed to use in random number generator for making initial guess of
+        training parameters. You should only set this if you want a script to
+        always generate the same "random" initial parameters. This primarily
+        useful for unit tests or benchmarking.
 
     .. note:: Dimensions of weight two dimensional arrays should be consistent
               with hiddenlayers.
@@ -94,9 +106,9 @@ class NeuralNetwork(Model):
     """
 
     def __init__(self, hiddenlayers=(5, 5), activation='tanh', weights=None,
-                 scalings=None, fprange=None, regressor=None, mode=None,
-                 lossfunction=None, retries=0, version=None, fortran=True,
-                 checkpoints=100):
+                 scalings=None, fprange=None, mode=None, version=None,
+                 regressor=None, lossfunction=None, fortran=True,
+                 checkpoints=100, randomseed=None, prescale=False):
 
         # Version check, particularly if restarting.
         compatibleversions = ['2015.12', ]
@@ -129,13 +141,14 @@ class NeuralNetwork(Model):
             raise NotImplementedError(_)
 
         self.regressor = regressor
-        self.parent = None  # Can hold a reference to main Amp instance.
+        self.parent = None  # Will hold a reference to main Amp instance.
         self.lossfunction = lossfunction
         self.fortran = fortran
         self.checkpoints = checkpoints
-        self.retries = retries
         if self.lossfunction is None:
             self.lossfunction = LossFunction()
+        self.randomseed = randomseed
+        self.prescale = prescale
 
     def fit(self,
             trainingimages,
@@ -165,14 +178,18 @@ class NeuralNetwork(Model):
             variables but skips the last line of starting the regressor.
         """
 
+        self.step = 0
+        self._log = log
+        self._load_from_checkpoints()  # if present; resume training
+
         # Set all parameters and report to logfile.
         self._parallel = parallel
-        self._log = log
 
         if self.regressor is None:
             self.regressor = Regressor()
 
         p = self.parameters
+
         tp = self.trainingparameters = Parameters()
         tp.images = trainingimages
         tp.descriptor = descriptor
@@ -204,13 +221,24 @@ class NeuralNetwork(Model):
 
         if p.weights is None:
             log('Initializing with random weights.')
-            self.randomize(scalings=False)
+            self.randomize(scalings=False, seed=self.randomseed)
         else:
             log('Initial weights already present.')
 
+        if p.scalings is None and self.prescale is True:
+            self.log('Finding good guesses for scaling intercepts...',
+                     tic='prescale')
+            self.prescale_intercepts(trainingimages)
+            self.log(' Atomic energies found:')
+            for element in self.parameters.scalings.keys():
+                self.log('{:2s}: {:14.4f}'.format(
+                    element, self.parameters.scalings[element]['intercept']))
+            self.log('...prescale complete.', toc='prescale')
+
         if p.scalings is None:
             log('Initializing with random scalings.')
-            self.randomize(weights=False, trainingimages=trainingimages)
+            self.randomize(weights=False, trainingimages=trainingimages,
+                           seed=self.randomseed)
         else:
             log('Initial scalings already present.')
 
@@ -218,18 +246,44 @@ class NeuralNetwork(Model):
             return
 
         # Regress the model.
-        result = False
-        tries = 0
-        while (result is False) and (tries <= self.retries):
-            log('Try {:d}/{:d}.'.format(tries, self.retries))
-            self.step = 0
-            if tries > 0:
-                self.randomize(trainingimages)
-            result = self.regressor.regress(model=self, log=log)
-            tries += 1
+        result = self.regressor.regress(model=self, log=log)
         return result  # True / False
 
-    def randomize(self, trainingimages=None, weights=True, scalings=True):
+    def prescale_intercepts(self, trainingimages):
+        """Calculates a reasonable guess for the per-atom energy, by regressing
+        a one-parameter-per-element model (per-atom energies), and uses this as
+        the scaling energy. This is a reasonably inefficient
+        residual-minimization technique, but doesn't typically bottleneck the
+        code. The slopes are not explicitly calculated, and take as w """
+
+        # This still assumes slope=1 is a goood guess.
+        # But in units of eV, it's not bad.
+
+        elements = self.parameters.fprange.keys()
+
+        def get_rmse_per_atom(scalings_list):
+            scalings = {element: scaling for
+                        (element, scaling) in zip(elements, scalings_list)}
+            calc = OffsetCalculator(scalings=scalings)
+            msea = 0.  # mean-squared (error per atom)
+            for image in trainingimages.values():
+                true_energy = image.get_potential_energy(
+                        apply_constraint=False)
+                predicted_energy = calc.get_potential_energy(image)
+                msea += ((true_energy - predicted_energy) / len(image))**2
+            return np.sqrt(msea)  # root-mean-squared (error per atom)
+
+        answer = fmin(get_rmse_per_atom, x0=[1.]*len(elements))
+        scaling_intercepts = {element: intercept for element, intercept in
+                              zip(elements, answer)}
+        p = self.parameters
+        p.scalings = {}
+        for element in elements:
+            p.scalings[element] = {'intercept': scaling_intercepts[element],
+                                   'slope': 1.}
+
+    def randomize(self, trainingimages=None, weights=True, scalings=True,
+                  seed=None):
         """Randomizes the model parameters (i.e., re-initializes them);
         this is typically used just before training.
 
@@ -242,6 +296,11 @@ class NeuralNetwork(Model):
             If False, do not randomize weights.
         scalings : bool
             If False, do not randomize scalings.
+        seed : None or int
+            Seed to use in random number generator for making initial guess of
+            training parameters. You should only set this if you want a script
+            to always generate the same "random" initial parameters. This
+            primarily useful for unit tests or benchmarking.
         """
         p = self.parameters
         if weights:
@@ -253,13 +312,13 @@ class NeuralNetwork(Model):
                 p.weights = get_random_weights(hiddenlayers=p.hiddenlayers,
                                                activation=p.activation,
                                                len_of_fps=len_of_fps,
-                                               )
+                                               seed=seed,)
         if scalings:
             if p.mode == 'image-centered':
                 raise NotImplementedError('Need to code.')
             elif p.mode == 'atom-centered':
-                p.scalings = get_random_scalings(trainingimages, p.activation,
-                                                 p.fprange.keys())
+                p.scalings = get_initial_scalings(trainingimages, p.activation,
+                                                  p.fprange.keys())
 
     @property
     def forcetraining(self):
@@ -314,7 +373,9 @@ class NeuralNetwork(Model):
         if self.step == 0:
             filename = make_filename(self.parent.label,
                                      '-initial-parameters.amp')
-            filename = self.parent.save(filename, overwrite=True)
+            if not os.path.exists(filename):
+                # If it exists, must be resuming from checkpoints.
+                filename = self.parent.save(filename)
         if self.checkpoints:
             if self.step % self.checkpoints == 0:
                 self._log('Saving checkpoint data.')
@@ -339,6 +400,28 @@ class NeuralNetwork(Model):
         else:
             return result['loss']
 
+
+    def _load_from_checkpoints(self):
+        """If checkpoints are present, this will load from them and therefore
+        resume a previous training run."""
+        # Check default checkpoint pattern.
+        filename = make_filename(self.parent.label, '-checkpoint.amp')
+        dirname = os.path.join(self.parent.label + '-checkpionts')
+        if os.path.exists(filename):
+            calc = Amp.load(filename, logging=False)
+        elif os.path.exists(dirname):
+            checkpoints = os.listdir(dirname)
+            last = sorted([int(_[:-4]) for _ in checkpoints])[-1]
+            filename = os.path.join(dirname, '{}.amp'.format(last))
+            calc = Amp.load(filename, logging=False)
+            self.step = last
+        else:
+            return  # No checkpoints present; run normally.
+        self._log('Found checkpoint file: {}.'.format(filename))
+        p = calc.model.parameters
+        self.parameters = p
+        self._log('Loaded last neural network parameters from checkpoint. '
+                  'Resuming training run.')
 
     def get_lossprime(self, vector):
         """
@@ -836,7 +919,7 @@ def calculate_ohat_D_delta(parameters, outputs, W):
 
 
 def get_random_weights(hiddenlayers, activation,
-                       len_of_fps=None, no_of_atoms=None,):
+                       len_of_fps=None, no_of_atoms=None, seed=None):
     """Generates random weight arrays from variables.
 
     hiddenlayers: dict
@@ -872,6 +955,11 @@ def get_random_weights(hiddenlayers, activation,
     no_of_atoms : int
         Number of atoms in atomic systems; used only in the case of no
         descriptor.
+    seed : None or int
+        Seed to use in random number generator for making initial guess of
+        training parameters. You should only set this if you want a script to
+        always generate the same "random" initial parameters. This primarily
+        useful for unit tests or benchmarking.
 
     Returns
     -------
@@ -879,6 +967,7 @@ def get_random_weights(hiddenlayers, activation,
         weights
     """
 
+    rs = np.random.RandomState(seed=seed)
     weight = {}
     nn_structure = {}
 
@@ -898,7 +987,7 @@ def get_random_weights(hiddenlayers, activation,
         epsilon = np.sqrt(6. / (nn_structure[0] +
                                 nn_structure[1]))
         normalized_arg_range = 2. * epsilon
-        weight[1] = np.random.random((3 * no_of_atoms + 1,
+        weight[1] = rs.random_sample((3 * no_of_atoms + 1,
                                       nn_structure[1])) * \
             normalized_arg_range - \
             normalized_arg_range / 2.
@@ -907,7 +996,7 @@ def get_random_weights(hiddenlayers, activation,
             epsilon = np.sqrt(6. / (nn_structure[layer + 1] +
                                     nn_structure[layer + 2]))
             normalized_arg_range = 2. * epsilon
-            weight[layer + 2] = np.random.random(
+            weight[layer + 2] = rs.random_sample(
                 (nn_structure[layer + 1] + 1,
                  nn_structure[layer + 2])) * \
                 normalized_arg_range - normalized_arg_range / 2.
@@ -916,7 +1005,7 @@ def get_random_weights(hiddenlayers, activation,
                                 nn_structure[-1]))
         normalized_arg_range = 2. * epsilon
         weight[len(list(nn_structure)) - 1] = \
-            np.random.random((nn_structure[-2] + 1, 1)) \
+            rs.random_sample((nn_structure[-2] + 1, 1)) \
             * normalized_arg_range - normalized_arg_range / 2.
 
         if False:  # This seemed to be setting all biases to zero?
@@ -945,7 +1034,7 @@ def get_random_weights(hiddenlayers, activation,
             epsilon = np.sqrt(6. / (nn_structure[element][0] +
                                     nn_structure[element][1]))
             normalized_arg_range = 2. * epsilon
-            weight[element][1] = (np.random.random(
+            weight[element][1] = (rs.random_sample(
                 (_len_of_fps + 1, nn_structure[element][1])) *
                 normalized_arg_range - normalized_arg_range / 2.)
             len_of_hiddenlayers = len(list(nn_structure[element])) - 3
@@ -953,7 +1042,7 @@ def get_random_weights(hiddenlayers, activation,
                 epsilon = np.sqrt(6. / (nn_structure[element][layer + 1] +
                                         nn_structure[element][layer + 2]))
                 normalized_arg_range = 2. * epsilon
-                weight[element][layer + 2] = np.random.random(
+                weight[element][layer + 2] = rs.random_sample(
                     (nn_structure[element][layer + 1] + 1,
                      nn_structure[element][layer + 2])) * \
                     normalized_arg_range - normalized_arg_range / 2.
@@ -962,7 +1051,7 @@ def get_random_weights(hiddenlayers, activation,
                                     nn_structure[element][-1]))
             normalized_arg_range = 2. * epsilon
             weight[element][len(list(nn_structure[element])) - 1] = \
-                np.random.random((nn_structure[element][-2] + 1, 1)) \
+                rs.random_sample((nn_structure[element][-2] + 1, 1)) \
                 * normalized_arg_range - normalized_arg_range / 2.
 
             if False:  # This seemed to be setting all biases to zero?
@@ -975,7 +1064,7 @@ def get_random_weights(hiddenlayers, activation,
     return weight
 
 
-def get_random_scalings(images, activation, elements=None):
+def get_initial_scalings(images, activation, elements=None, seed=None):
     """Generates initial scaling matrices, such that the range of activation is
     scaled to the range of actual energies.
 
@@ -1061,7 +1150,6 @@ def get_random_scalings(images, activation, elements=None):
 
 
 class Raveler:
-
     """Class to ravel and unravel variable values into a single vector.
 
     This is used for feeding into the optimizer. Feed in a list of dictionaries
@@ -1222,6 +1310,7 @@ class NodePlot:
         """Accumulates the data for the symbol."""
         data = self.data
         layerkeys = list(output.keys())  # Correspond to layers.
+        layerkeys.sort()
 
         if symbol not in data:
             # Create headers, structure.
@@ -1244,3 +1333,20 @@ class NodePlot:
         """Converts the data table into a numpy array."""
         for symbol in self.data:
             self.data[symbol]['table'] = np.array(self.data[symbol]['table'])
+
+
+class OffsetCalculator:
+    """A calculator in which energy is only a sum of one-body atomic terms,
+    with an energy per element type. This is used to calculate the scaling
+    parameters."""
+
+    def __init__(self, scalings):
+        """Scalings is a dictionary [element:energy]."""
+        self.scalings = scalings
+
+    def get_potential_energy(self, atoms):
+        """Calculate a crude potential energy."""
+        energy = 0.
+        for atom in atoms:
+            energy += self.scalings[atom.symbol]
+        return energy